User Friendly Update
Reworked the Main.py codes to be more user friendly with the template files. 

template.txt and emojilist4.csv must now be in the Code folder of RITHM

template.txt can now be relatively as sloppy as you want. the commands can be in any order (as long as the KWs are still last) and the command words that you use can vary so long as there is a colon following it (example "start:" and "begin:" both work as well as "end: and "finish:")